+++ app/src/main/java/woowacourse/shopping/application/ShoppingApplication.kt
@@ -1,9 +1,9 @@
 package woowacourse.shopping.application
 
-import woowacourse.shopping.di.DIApplication
-import woowacourse.shopping.di.DefaultActivityModule
-import woowacourse.shopping.di.DefaultApplicationModule
-import woowacourse.shopping.di.DefaultViewModelModule
+import woowacourse.shopping.di.application.DIApplication
+import woowacourse.shopping.di.module.DefaultActivityModule
+import woowacourse.shopping.di.module.DefaultApplicationModule
+import woowacourse.shopping.di.module.DefaultViewModelModule
 
 class ShoppingApplication : DIApplication(
     DefaultApplicationModule::class.java,
