+++ app/src/main/java/woowacourse/shopping/di/annotation/SAnnotation.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.di
+package woowacourse.shopping.di.annotation
 
 @Qualifier
 annotation class RoomDB
