+++ app/src/main/java/woowacourse/shopping/di/module/DefaultApplicationModule.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.di
+package woowacourse.shopping.di.module
 
 import android.content.Context
 import woowacourse.shopping.data.CartProductDao
@@ -6,6 +6,8 @@ import woowacourse.shopping.data.CartRepository
 import woowacourse.shopping.data.InMemoryCartRepository
 import woowacourse.shopping.data.RoomCartRepository
 import woowacourse.shopping.data.ShoppingDatabase
+import woowacourse.shopping.di.annotation.InMemory
+import woowacourse.shopping.di.annotation.RoomDB
 
 class DefaultApplicationModule(
     applicationContext: Context,
