+++ app/src/main/java/woowacourse/shopping/di/module/DefaultViewModelModule.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.di
+package woowacourse.shopping.di.module
 
 import woowacourse.shopping.data.DefaultProductRepository
 import woowacourse.shopping.data.ProductRepository
