+++ di/src/main/java/woowacourse/shopping/di/annotation/SAnnotation.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.di
+package woowacourse.shopping.di.annotation
 
 @Target(AnnotationTarget.PROPERTY, AnnotationTarget.FIELD)
 annotation class Inject
