+++ di/src/main/java/woowacourse/shopping/di/module/ViewModelModule.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.di
+package woowacourse.shopping.di.module
 
 import androidx.lifecycle.ViewModel
